Fix 1.9 types
In 1.9, writing varints and such return the packetbuffer where previously they returned void.  Thus, the existing tests did not find the method, and then the position method was assumed (it's used for packetbuffer methods that don't match anything else, since there is no other good way to identify it).  The fix is to test if a method is obfuscated by the length, thus preventing ambiguity between writeInt and writeVarInt.