Fix entities and language toppings to work with 16w32a
Mojang significantly changed the way entities were registered in 16w32a, necessitating separate logic for reading them between versions.  The change to the language topping is because language files are now lowercase; nothing major.  The sounds topping change makes it so that the 1.11 sound index is used, though right now that may still be the wrong index for earlier versions; future work to use the appropriate index is necessary.