Fix tile entity topping's block identification
Namespaces are now used there, and there's two maps (one for old IDs and one for new ones).  Since the regular tile entity code doesn't use namespaces explicitly, we need to strip the namespace to match it.  We also need to make sure to only use the new ID map.