Fix unwanted method call paren striping
Previously, results like '((Integer)var2.get(var6)).intValue()' would be changed to '(Integer)var2.get(var6)).intValue(' as they started and ended with a paren.  My fix is to not remove the paren if there is an opening paren right before.  This might not work perfectly for methods with parameters, though (but I don't think that case is very common)