Fix blockstates failing if sounds failed
For performance, I skip the sounds list; however, this might not be identified properly if the sounds topping didn't run (or failed to download).  Both changes fix this issue in separate ways, one where failing to download will still find the class, and the other where if it still isn't identified (e.g. if sounds was completely skipped) we just run through the class anyways.