Fix identification of entity list in 20w30a
Refactoring in that snapshot moved the disconnect.lost string constant in NHPC into a static initializer, which put it at the end of the constant list.  NHPC also has the Skipping Entity with id constant, which means that it ended up matching as the entity list instead (since that was now before disconnect.lost).  Then the actual entity list also matched, which threw an exception.

The fix is to make Skipping Entity with id a lower priority constant.