Fix blocks topping generating invalid json
The various return object() statements cause problems if they end up in the result, since object() can't be serialized.  In this case, it's because candles have a light level dependent on a block state, and the lambda for that is stored as a field (instead of being in the block list).  Work around by returning None for that case (same as if the lambda were in the list, but failed to evaluate)