Fix block entities topping
Two changes.  First, there are now 3 parameters in the block entity constructor (the first is (still) the type, the second is the position, and the third is some kind of serialization thing).  Second, the provider now has two methods (one which is defaulted); we care about the one that actually creates block entities though.